// src/solratis/issue/api/DefaultApi.ts
export class DefaultApi {
    protected basePath = 'https://travelapihk.solartis.net/DroolsV4_2/DroolsService';
    public defaultHeaders: Array<string> = [];
    public configuration: Configuration = new Configuration();

    constructor(basePath?: string, configuration?: Configuration) {
        if (basePath) {
            this.basePath = basePath;
        }
        if (configuration) {
            this.configuration = configuration;
        }
    }

    private extendObj<T1, T2 extends T1>(objA: T2, objB: T2): T1|T2 {
        for (let key in objB) {
            if (objB.hasOwnProperty(key)) {
                objA[key] = objB[key];
            }
        }
        return objA;
    }

    /**
     * Policy Issuance for Leisure Travel
     * @param Token Authentication Token
     * @param ContentType Content Type
     * @param EventName EventName
     * @param Request For request
     */
    public policyissuance(Token: string, ContentType: string, EventName: string, Request: models.Request): JQueryPromise<{ response: JQueryXHR; body: models.SuccessResponse;  }> {
        let localVarPath = this.basePath + '/FireEventV2';

        let queryParameters: any = {};
        let headerParams: any = {};
        // verify required parameter 'Token' is not null or undefined
        if (Token === null || Token === undefined) {
            throw new Error('Required parameter Token was null or undefined when calling policyissuance.');
        }

        // verify required parameter 'ContentType' is not null or undefined
        if (ContentType === null || ContentType === undefined) {
            throw new Error('Required parameter ContentType was null or undefined when calling policyissuance.');
        }

        // verify required parameter 'EventName' is not null or undefined
        if (EventName === null || EventName === undefined) {
            throw new Error('Required parameter EventName was null or undefined when calling policyissuance.');
        }

        // verify required parameter 'Request' is not null or undefined
        if (Request === null || Request === undefined) {
            throw new Error('Required parameter Request was null or undefined when calling policyissuance.');
        }


        localVarPath = localVarPath + "?" + $.param(queryParameters);
        headerParams['Token'] = String(Token);

        headerParams['Content-Type'] = String(ContentType);

        headerParams['EventName'] = String(EventName);

        // to determine the Content-Type header
        let consumes: string[] = [
            'application/json'
        ];

        // to determine the Accept header
        let produces: string[] = [
            'application/json'
        ];


        headerParams['Content-Type'] = 'application/json';

        let requestOptions: JQueryAjaxSettings = {
            url: localVarPath,
            type: 'POST',
            headers: headerParams,
            processData: false
        };

        requestOptions.data = JSON.stringify(Request);
        if (headerParams['Content-Type']) {
            requestOptions.contentType = headerParams['Content-Type'];
        }

        let dfd = $.Deferred();
        $.ajax(requestOptions).then(
            (data: models.SuccessResponse, textStatus: string, jqXHR: JQueryXHR) =>
                dfd.resolve(jqXHR, data),
            (xhr: JQueryXHR, textStatus: string, errorThrown: string) =>
                dfd.reject(xhr, errorThrown)
        );
        return dfd.promise();
    }

}